Working overlays!
An overlay is a sub-room shown on top of the main room, and allows the
player to do things like investigate the inside of cabinets.  When in an
overlay the watch is going to be hidden, and interactions are only
available for clickables within the overlay itself.

While overlays existed last time, this commit irons out issues i was
having with dismissing them, interacting with clickables within them,
and click events being handled by clickables in the room behind them.  I
believe this solves all of those issues.