Fix camera between days, fix bug in collectable
When starting the level or transitioning between days, the camera should
snap to an initial posiiton; that's handled now.

Added two more collectables and noticed that their collision shapes were
off; turns out I needed to make that local to the Collectable.tscn scene
so they didn't all share and resize the same shape - otherwise the last
shape to load would define the shape of all Collectables in the level.